Update README.md
Overview

For this assignment, you will be implementing parts of a class to allow for operator overloads, as well as to remove some aspects of the special member functions.

Download the starter code here.

Part 1: Make a profile

The newest Stanford social media startup is Treebook, and you’re a founding member of the team! To get the product off the ground and compete with an unnamed, completely legally unaffiliated app from Harvard, you’ve been assigned the task of implementing user profiles.

You will be writing the User class into the user.cpp file. The header user.h has been provided and started for you, but you will need to add special member functions to it.